Change cost to microdollars

+++ backend/config/settings/base.py
@@ -450,7 +450,7 @@ PIPELINE_RUN_DEFAULT_TIMEOUT = os.environ.get("PIPELINE_RUN_DEFAULT_TIMEOUT", 14
 PIPELINE_RUN_MAX_TIMEOUT = os.environ.get("PIPELINE_RUN_MAX_TIMEOUT", 43200)
 
 # AI Assistant monthly limit in USD
-ASSISTANT_MONTHLY_LIMIT = float(os.environ.get("ASSISTANT_MONTHLY_LIMIT", 200))
+ASSISTANT_MONTHLY_LIMIT = int(os.environ.get("ASSISTANT_MONTHLY_LIMIT", 200))
 
 # Two Factor Authentication
 OTP_EMAIL_BODY_TEMPLATE_PATH = "user_management/token.txt"
